// src/app/challenge-failer/models/class.ts
export enum Class {
  Feca = "Feca",
  Osamodas = "Osamodas",
  Enutrof = "Enutrof",
  Sram = "Sram",
  Xelor = "Xelor",
  Ecaflip = "Ecaflip",
  Eniripsa = "Eniripsa",
  Iop = "Iop",
  Cra = "Cra",
  Sadida = "Sadida",
  Sacrieur = "Sacrieur",
  Pandawa = "Pandawa",
  Roublard = "Roublard",
  Zobal = "Zobal",
  Steamer = "Steamer",
  Eliotrope = "Eliotrope",
  Huppermage = "Huppermage",
  Ouginak = "Ouginak",
  Forgelance = "Forgelance",
}